replacing SF tableNameComponents with PG's version
